CPR-1159 move LIBRA event into correct package
